requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
pyTelegramBotAPI==4.13.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.4
cffi==1.15.1
cryptography==41.0.4
dill==0.3.7
lxml==4.9.3
multiprocess==0.70.15
pathos==0.3.1
pox==0.3.3
ppft==1.7.6.7
pycparser==2.21
PyExecJS==1.5.1
six==1.16.0
tqdm==4.66.1
langdetect==1.0.9
chardet==3.0.4
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
rfc3986==1.5.0
sniffio==1.3.0
